Shapiro-Wilk Normality Test

Performs the Shapiro–Wilk normality test, which assesses whether a sample originates from a normally distributed population using a regression-based correlation method.

Usage

Shapiro_Wilk_test(
  x,
  alpha = 0.05,
  method = c("SWR", "SF", "SW"),
  silent = FALSE,
  summary = TRUE,
  misc = FALSE,
  resampling = TRUE
)

Arguments

x

A numeric vector.

alpha

Significance threshold (default: 0.05).

method

Character (default: "SWR"). Use which modification of the test? Available options are c("SWR", "SF", "SW").

silent

Logical (default: FALSE). If FALSE, print out the results.

summary

Logical (default: TRUE). Produce a summary table.

misc

Logical (default: FALSE). Output other unimportant parameters.

resampling

Logical (default: TRUE). If TRUE, unlock the sample size limitation of the test by using sample resampling method.

Value

A list.

Details

method

"SW": Shapiro–Wilk test, the original normality test proposed by Shapiro and Wilk (1965). Applicable only for sample sizes 3 <= n <= 50.

"SF": Shapiro–Francia test, proposed by Shapiro and Francia (1972) and subsequently simplified and extended by Royston (1993). Applicable only for sample sizes 5 <= n <= 5000.

"SWR": Shapiro–Wilk test with Royston's (1992) modifications for approximating the null distribution and extending the test to larger sample sizes. Applicable only for sample sizes 3 <= n <= 5000.

References

Shapiro, S.S., Wilk, M.B., 1965. An Analysis of Variance Test for Normality (Complete Samples). Biometrika 52, 591–611. https://doi.org/10.2307/2333709

Shapiro, S.S., Francia, R.S., 1972. An Approximate Analysis of Variance Test for Normality. J. Am. Stat. Assoc. 67, 215–216. https://doi.org/10.1080/01621459.1972.10481232

Royston, P., 1993. A pocket-calculator algorithm for the Shapiro–Francia test for non-normality: an application to medicine. Stat. Med. 12, 181–184. https://doi.org/10.1002/sim.4780120209

Royston, P., 1992. Approximating the Shapiro–Wilk W-test for non-normality. Stat. Comput. 2, 117–119. https://doi.org/10.1007/BF01891203

Examples

sw <- Shapiro_Wilk_test(rnorm(20), method = "SW")
#> 
#> ---------------------------------
#>  Shapiro-Wilk (W) normality test  
#> 
#>  Statistic (W) = 0.9824
#>  p-value = 0.9428 
#> ---------------------------------
print(sw$summary)
#>                  alpha statistic     pval signif standard_value critical_value
#> Shapiro-Wilk (W)  0.05 0.9824243 0.942804     ns      0.9824243          0.905
#>                  SE CI_lower CI_upper  N        AVG          MED      MIN
#> Shapiro-Wilk (W) NA       NA       NA 20 0.08252507 -0.002685395 -2.44668
#>                       MAX       SD
#> Shapiro-Wilk (W) 2.648932 1.251687
sf <- Shapiro_Wilk_test(rnorm(100) ^ 2, method = "SF")
#> 
#> -------------------------------------
#>  Shapiro-Francia (W') normality test  
#> 
#>  Statistic (W) = 0.8125
#>  p-value = 0 
#> -------------------------------------
print(sf$summary)
#>                      alpha statistic         pval signif standard_value
#> Shapiro-Francia (W')  0.05 0.8124924 1.028573e-08    ***       5.607125
#>                      critical_value SE CI_lower CI_upper   N     AVG       MED
#> Shapiro-Francia (W')       1.644854 NA       NA       NA 100 1.02848 0.5766757
#>                               MIN      MAX       SD
#> Shapiro-Francia (W') 9.771503e-06 4.741837 1.195992
swr <- Shapiro_Wilk_test(rnorm(1e6), method = "SWR")
#> Resampling: 1/200
#> Resampling: 2/200
#> Resampling: 3/200
#> Resampling: 4/200
#> Resampling: 5/200
#> Resampling: 6/200
#> Resampling: 7/200
#> Resampling: 8/200
#> Resampling: 9/200
#> Resampling: 10/200
#> Resampling: 11/200
#> Resampling: 12/200
#> Resampling: 13/200
#> Resampling: 14/200
#> Resampling: 15/200
#> Resampling: 16/200
#> Resampling: 17/200
#> Resampling: 18/200
#> Resampling: 19/200
#> Resampling: 20/200
#> Resampling: 21/200
#> Resampling: 22/200
#> Resampling: 23/200
#> Resampling: 24/200
#> Resampling: 25/200
#> Resampling: 26/200
#> Resampling: 27/200
#> Resampling: 28/200
#> Resampling: 29/200
#> Resampling: 30/200
#> Resampling: 31/200
#> Resampling: 32/200
#> Resampling: 33/200
#> Resampling: 34/200
#> Resampling: 35/200
#> Resampling: 36/200
#> Resampling: 37/200
#> Resampling: 38/200
#> Resampling: 39/200
#> Resampling: 40/200
#> Resampling: 41/200
#> Resampling: 42/200
#> Resampling: 43/200
#> Resampling: 44/200
#> Resampling: 45/200
#> Resampling: 46/200
#> Resampling: 47/200
#> Resampling: 48/200
#> Resampling: 49/200
#> Resampling: 50/200
#> Resampling: 51/200
#> Resampling: 52/200
#> Resampling: 53/200
#> Resampling: 54/200
#> Resampling: 55/200
#> Resampling: 56/200
#> Resampling: 57/200
#> Resampling: 58/200
#> Resampling: 59/200
#> Resampling: 60/200
#> Resampling: 61/200
#> Resampling: 62/200
#> Resampling: 63/200
#> Resampling: 64/200
#> Resampling: 65/200
#> Resampling: 66/200
#> Resampling: 67/200
#> Resampling: 68/200
#> Resampling: 69/200
#> Resampling: 70/200
#> Resampling: 71/200
#> Resampling: 72/200
#> Resampling: 73/200
#> Resampling: 74/200
#> Resampling: 75/200
#> Resampling: 76/200
#> Resampling: 77/200
#> Resampling: 78/200
#> Resampling: 79/200
#> Resampling: 80/200
#> Resampling: 81/200
#> Resampling: 82/200
#> Resampling: 83/200
#> Resampling: 84/200
#> Resampling: 85/200
#> Resampling: 86/200
#> Resampling: 87/200
#> Resampling: 88/200
#> Resampling: 89/200
#> Resampling: 90/200
#> Resampling: 91/200
#> Resampling: 92/200
#> Resampling: 93/200
#> Resampling: 94/200
#> Resampling: 95/200
#> Resampling: 96/200
#> Resampling: 97/200
#> Resampling: 98/200
#> Resampling: 99/200
#> Resampling: 100/200
#> Resampling: 101/200
#> Resampling: 102/200
#> Resampling: 103/200
#> Resampling: 104/200
#> Resampling: 105/200
#> Resampling: 106/200
#> Resampling: 107/200
#> Resampling: 108/200
#> Resampling: 109/200
#> Resampling: 110/200
#> Resampling: 111/200
#> Resampling: 112/200
#> Resampling: 113/200
#> Resampling: 114/200
#> Resampling: 115/200
#> Resampling: 116/200
#> Resampling: 117/200
#> Resampling: 118/200
#> Resampling: 119/200
#> Resampling: 120/200
#> Resampling: 121/200
#> Resampling: 122/200
#> Resampling: 123/200
#> Resampling: 124/200
#> Resampling: 125/200
#> Resampling: 126/200
#> Resampling: 127/200
#> Resampling: 128/200
#> Resampling: 129/200
#> Resampling: 130/200
#> Resampling: 131/200
#> Resampling: 132/200
#> Resampling: 133/200
#> Resampling: 134/200
#> Resampling: 135/200
#> Resampling: 136/200
#> Resampling: 137/200
#> Resampling: 138/200
#> Resampling: 139/200
#> Resampling: 140/200
#> Resampling: 141/200
#> Resampling: 142/200
#> Resampling: 143/200
#> Resampling: 144/200
#> Resampling: 145/200
#> Resampling: 146/200
#> Resampling: 147/200
#> Resampling: 148/200
#> Resampling: 149/200
#> Resampling: 150/200
#> Resampling: 151/200
#> Resampling: 152/200
#> Resampling: 153/200
#> Resampling: 154/200
#> Resampling: 155/200
#> Resampling: 156/200
#> Resampling: 157/200
#> Resampling: 158/200
#> Resampling: 159/200
#> Resampling: 160/200
#> Resampling: 161/200
#> Resampling: 162/200
#> Resampling: 163/200
#> Resampling: 164/200
#> Resampling: 165/200
#> Resampling: 166/200
#> Resampling: 167/200
#> Resampling: 168/200
#> Resampling: 169/200
#> Resampling: 170/200
#> Resampling: 171/200
#> Resampling: 172/200
#> Resampling: 173/200
#> Resampling: 174/200
#> Resampling: 175/200
#> Resampling: 176/200
#> Resampling: 177/200
#> Resampling: 178/200
#> Resampling: 179/200
#> Resampling: 180/200
#> Resampling: 181/200
#> Resampling: 182/200
#> Resampling: 183/200
#> Resampling: 184/200
#> Resampling: 185/200
#> Resampling: 186/200
#> Resampling: 187/200
#> Resampling: 188/200
#> Resampling: 189/200
#> Resampling: 190/200
#> Resampling: 191/200
#> Resampling: 192/200
#> Resampling: 193/200
#> Resampling: 194/200
#> Resampling: 195/200
#> Resampling: 196/200
#> Resampling: 197/200
#> Resampling: 198/200
#> Resampling: 199/200
#> Resampling: 200/200
#> 
#> -----------------------------------------
#>  Shapiro-Wilk-Royston (w) normality test  
#> 
#>  Statistic (W) = 0.9999
#>  p-value = 1 
#> -----------------------------------------
print(swr$summary)
#>              alpha statistic      pval signif standard_value critical_value SE
#> resample_1    0.05 0.9995387 0.2791006     ns      0.5855154       1.644854 NA
#> resample_2    0.05 0.9995685 0.3408326     ns      0.4101920       1.644854 NA
#> resample_3    0.05 0.9998277 0.9771931     ns     -1.9989505       1.644854 NA
#> resample_4    0.05 0.9998567 0.9934876     ns     -2.4830928       1.644854 NA
#> resample_5    0.05 0.9998624 0.9951867     ns     -2.5889590       1.644854 NA
#> resample_6    0.05 0.9998759 0.9978834     ns     -2.8602331       1.644854 NA
#> resample_7    0.05 0.9998886 0.9991665     ns     -3.1439092       1.644854 NA
#> resample_8    0.05 0.9999040 0.9997957     ns     -3.5345298       1.644854 NA
#> resample_9    0.05 0.9999058 0.9998304     ns     -3.5833731       1.644854 NA
#> resample_10   0.05 0.9999206 0.9999724     ns     -4.0326323       1.644854 NA
#> resample_11   0.05 0.9999244 0.9999840     ns     -4.1594423       1.644854 NA
#> resample_12   0.05 0.9999279 0.9999908     ns     -4.2829316       1.644854 NA
#> resample_13   0.05 0.9999324 0.9999958     ns     -4.4539409       1.644854 NA
#> resample_14   0.05 0.9999336 0.9999966     ns     -4.4997920       1.644854 NA
#> resample_15   0.05 0.9999341 0.9999969     ns     -4.5206844       1.644854 NA
#> resample_16   0.05 0.9999354 0.9999976     ns     -4.5716748       1.644854 NA
#> resample_17   0.05 0.9999434 0.9999996     ns     -4.9207919       1.644854 NA
#> resample_18   0.05 0.9999449 0.9999997     ns     -4.9914529       1.644854 NA
#> resample_19   0.05 0.9999452 0.9999997     ns     -5.0067949       1.644854 NA
#> resample_20   0.05 0.9999496 0.9999999     ns     -5.2244940       1.644854 NA
#> resample_21   0.05 0.9999507 0.9999999     ns     -5.2845575       1.644854 NA
#> resample_22   0.05 0.9999528 1.0000000     ns     -5.3982379       1.644854 NA
#> resample_23   0.05 0.9999536 1.0000000     ns     -5.4383237       1.644854 NA
#> resample_24   0.05 0.9999564 1.0000000     ns     -5.6044307       1.644854 NA
#> resample_25   0.05 0.9999581 1.0000000     ns     -5.7070256       1.644854 NA
#> resample_26   0.05 0.9999604 1.0000000     ns     -5.8554061       1.644854 NA
#> resample_27   0.05 0.9999608 1.0000000     ns     -5.8842998       1.644854 NA
#> resample_28   0.05 0.9999614 1.0000000     ns     -5.9222647       1.644854 NA
#> resample_29   0.05 0.9999620 1.0000000     ns     -5.9633730       1.644854 NA
#> resample_30   0.05 0.9999625 1.0000000     ns     -6.0001528       1.644854 NA
#> resample_31   0.05 0.9999629 1.0000000     ns     -6.0305740       1.644854 NA
#> resample_32   0.05 0.9999634 1.0000000     ns     -6.0609156       1.644854 NA
#> resample_33   0.05 0.9999641 1.0000000     ns     -6.1144337       1.644854 NA
#> resample_34   0.05 0.9999655 1.0000000     ns     -6.2215813       1.644854 NA
#> resample_35   0.05 0.9999663 1.0000000     ns     -6.2806807       1.644854 NA
#> resample_36   0.05 0.9999667 1.0000000     ns     -6.3097267       1.644854 NA
#> resample_37   0.05 0.9999671 1.0000000     ns     -6.3455992       1.644854 NA
#> resample_38   0.05 0.9999676 1.0000000     ns     -6.3820208       1.644854 NA
#> resample_39   0.05 0.9999682 1.0000000     ns     -6.4364820       1.644854 NA
#> resample_40   0.05 0.9999696 1.0000000     ns     -6.5492349       1.644854 NA
#> resample_41   0.05 0.9999699 1.0000000     ns     -6.5774674       1.644854 NA
#> resample_42   0.05 0.9999703 1.0000000     ns     -6.6115765       1.644854 NA
#> resample_43   0.05 0.9999708 1.0000000     ns     -6.6592117       1.644854 NA
#> resample_44   0.05 0.9999713 1.0000000     ns     -6.7053466       1.644854 NA
#> resample_45   0.05 0.9999725 1.0000000     ns     -6.8109568       1.644854 NA
#> resample_46   0.05 0.9999734 1.0000000     ns     -6.9011942       1.644854 NA
#> resample_47   0.05 0.9999737 1.0000000     ns     -6.9307431       1.644854 NA
#> resample_48   0.05 0.9999744 1.0000000     ns     -6.9992303       1.644854 NA
#> resample_49   0.05 0.9999747 1.0000000     ns     -7.0297885       1.644854 NA
#> resample_50   0.05 0.9999749 1.0000000     ns     -7.0552455       1.644854 NA
#> resample_51   0.05 0.9999751 1.0000000     ns     -7.0757332       1.644854 NA
#> resample_52   0.05 0.9999753 1.0000000     ns     -7.0977561       1.644854 NA
#> resample_53   0.05 0.9999760 1.0000000     ns     -7.1699565       1.644854 NA
#> resample_54   0.05 0.9999763 1.0000000     ns     -7.1983952       1.644854 NA
#> resample_55   0.05 0.9999766 1.0000000     ns     -7.2368755       1.644854 NA
#> resample_56   0.05 0.9999768 1.0000000     ns     -7.2581975       1.644854 NA
#> resample_57   0.05 0.9999775 1.0000000     ns     -7.3427528       1.644854 NA
#> resample_58   0.05 0.9999778 1.0000000     ns     -7.3703544       1.644854 NA
#> resample_59   0.05 0.9999778 1.0000000     ns     -7.3753715       1.644854 NA
#> resample_60   0.05 0.9999780 1.0000000     ns     -7.3951531       1.644854 NA
#> resample_61   0.05 0.9999781 1.0000000     ns     -7.4116776       1.644854 NA
#> resample_62   0.05 0.9999786 1.0000000     ns     -7.4689645       1.644854 NA
#> resample_63   0.05 0.9999788 1.0000000     ns     -7.5008634       1.644854 NA
#> resample_64   0.05 0.9999789 1.0000000     ns     -7.5087435       1.644854 NA
#> resample_65   0.05 0.9999794 1.0000000     ns     -7.5736522       1.644854 NA
#> resample_66   0.05 0.9999797 1.0000000     ns     -7.6067658       1.644854 NA
#> resample_67   0.05 0.9999799 1.0000000     ns     -7.6304422       1.644854 NA
#> resample_68   0.05 0.9999802 1.0000000     ns     -7.6718129       1.644854 NA
#> resample_69   0.05 0.9999803 1.0000000     ns     -7.6891422       1.644854 NA
#> resample_70   0.05 0.9999804 1.0000000     ns     -7.7022007       1.644854 NA
#> resample_71   0.05 0.9999811 1.0000000     ns     -7.7940750       1.644854 NA
#> resample_72   0.05 0.9999812 1.0000000     ns     -7.8118334       1.644854 NA
#> resample_73   0.05 0.9999816 1.0000000     ns     -7.8693011       1.644854 NA
#> resample_74   0.05 0.9999821 1.0000000     ns     -7.9394230       1.644854 NA
#> resample_75   0.05 0.9999822 1.0000000     ns     -7.9567230       1.644854 NA
#> resample_76   0.05 0.9999823 1.0000000     ns     -7.9722975       1.644854 NA
#> resample_77   0.05 0.9999825 1.0000000     ns     -8.0005809       1.644854 NA
#> resample_78   0.05 0.9999828 1.0000000     ns     -8.0395735       1.644854 NA
#> resample_79   0.05 0.9999825 1.0000000     ns     -8.0058905       1.644854 NA
#> resample_80   0.05 0.9999827 1.0000000     ns     -8.0232626       1.644854 NA
#> resample_81   0.05 0.9999828 1.0000000     ns     -8.0504143       1.644854 NA
#> resample_82   0.05 0.9999831 1.0000000     ns     -8.0883931       1.644854 NA
#> resample_83   0.05 0.9999832 1.0000000     ns     -8.1010636       1.644854 NA
#> resample_84   0.05 0.9999833 1.0000000     ns     -8.1170683       1.644854 NA
#> resample_85   0.05 0.9999835 1.0000000     ns     -8.1499355       1.644854 NA
#> resample_86   0.05 0.9999836 1.0000000     ns     -8.1697252       1.644854 NA
#> resample_87   0.05 0.9999838 1.0000000     ns     -8.2102221       1.644854 NA
#> resample_88   0.05 0.9999840 1.0000000     ns     -8.2393180       1.644854 NA
#> resample_89   0.05 0.9999841 1.0000000     ns     -8.2488018       1.644854 NA
#> resample_90   0.05 0.9999846 1.0000000     ns     -8.3330192       1.644854 NA
#> resample_91   0.05 0.9999847 1.0000000     ns     -8.3565434       1.644854 NA
#> resample_92   0.05 0.9999851 1.0000000     ns     -8.4146403       1.644854 NA
#> resample_93   0.05 0.9999851 1.0000000     ns     -8.4233004       1.644854 NA
#> resample_94   0.05 0.9999846 1.0000000     ns     -8.3409408       1.644854 NA
#> resample_95   0.05 0.9999846 1.0000000     ns     -8.3347351       1.644854 NA
#> resample_96   0.05 0.9999846 1.0000000     ns     -8.3340577       1.644854 NA
#> resample_97   0.05 0.9999849 1.0000000     ns     -8.3866562       1.644854 NA
#> resample_98   0.05 0.9999853 1.0000000     ns     -8.4560194       1.644854 NA
#> resample_99   0.05 0.9999855 1.0000000     ns     -8.4970848       1.644854 NA
#> resample_100  0.05 0.9999855 1.0000000     ns     -8.4990990       1.644854 NA
#> resample_101  0.05 0.9999855 1.0000000     ns     -8.4980764       1.644854 NA
#> resample_102  0.05 0.9999853 1.0000000     ns     -8.4624526       1.644854 NA
#> resample_103  0.05 0.9999851 1.0000000     ns     -8.4221130       1.644854 NA
#> resample_104  0.05 0.9999847 1.0000000     ns     -8.3480432       1.644854 NA
#> resample_105  0.05 0.9999843 1.0000000     ns     -8.2815025       1.644854 NA
#> resample_106  0.05 0.9999843 1.0000000     ns     -8.2803234       1.644854 NA
#> resample_107  0.05 0.9999844 1.0000000     ns     -8.3070110       1.644854 NA
#> resample_108  0.05 0.9999845 1.0000000     ns     -8.3110588       1.644854 NA
#> resample_109  0.05 0.9999846 1.0000000     ns     -8.3392524       1.644854 NA
#> resample_110  0.05 0.9999846 1.0000000     ns     -8.3285056       1.644854 NA
#> resample_111  0.05 0.9999847 1.0000000     ns     -8.3528074       1.644854 NA
#> resample_112  0.05 0.9999847 1.0000000     ns     -8.3479228       1.644854 NA
#> resample_113  0.05 0.9999845 1.0000000     ns     -8.3210863       1.644854 NA
#> resample_114  0.05 0.9999839 1.0000000     ns     -8.2104669       1.644854 NA
#> resample_115  0.05 0.9999840 1.0000000     ns     -8.2350208       1.644854 NA
#> resample_116  0.05 0.9999845 1.0000000     ns     -8.3110557       1.644854 NA
#> resample_117  0.05 0.9999846 1.0000000     ns     -8.3285297       1.644854 NA
#> resample_118  0.05 0.9999848 1.0000000     ns     -8.3704685       1.644854 NA
#> resample_119  0.05 0.9999847 1.0000000     ns     -8.3565590       1.644854 NA
#> resample_120  0.05 0.9999845 1.0000000     ns     -8.3159653       1.644854 NA
#> resample_121  0.05 0.9999846 1.0000000     ns     -8.3357233       1.644854 NA
#> resample_122  0.05 0.9999847 1.0000000     ns     -8.3539977       1.644854 NA
#> resample_123  0.05 0.9999847 1.0000000     ns     -8.3465613       1.644854 NA
#> resample_124  0.05 0.9999847 1.0000000     ns     -8.3467076       1.644854 NA
#> resample_125  0.05 0.9999847 1.0000000     ns     -8.3462868       1.644854 NA
#> resample_126  0.05 0.9999848 1.0000000     ns     -8.3772582       1.644854 NA
#> resample_127  0.05 0.9999850 1.0000000     ns     -8.4048743       1.644854 NA
#> resample_128  0.05 0.9999851 1.0000000     ns     -8.4165026       1.644854 NA
#> resample_129  0.05 0.9999848 1.0000000     ns     -8.3651981       1.644854 NA
#> resample_130  0.05 0.9999845 1.0000000     ns     -8.3110267       1.644854 NA
#> resample_131  0.05 0.9999846 1.0000000     ns     -8.3314279       1.644854 NA
#> resample_132  0.05 0.9999846 1.0000000     ns     -8.3289015       1.644854 NA
#> resample_133  0.05 0.9999845 1.0000000     ns     -8.3122822       1.644854 NA
#> resample_134  0.05 0.9999844 1.0000000     ns     -8.3017014       1.644854 NA
#> resample_135  0.05 0.9999843 1.0000000     ns     -8.2800061       1.644854 NA
#> resample_136  0.05 0.9999843 1.0000000     ns     -8.2850515       1.644854 NA
#> resample_137  0.05 0.9999844 1.0000000     ns     -8.2948695       1.644854 NA
#> resample_138  0.05 0.9999844 1.0000000     ns     -8.2968700       1.644854 NA
#> resample_139  0.05 0.9999842 1.0000000     ns     -8.2653292       1.644854 NA
#> resample_140  0.05 0.9999841 1.0000000     ns     -8.2561493       1.644854 NA
#> resample_141  0.05 0.9999840 1.0000000     ns     -8.2324164       1.644854 NA
#> resample_142  0.05 0.9999839 1.0000000     ns     -8.2132226       1.644854 NA
#> resample_143  0.05 0.9999834 1.0000000     ns     -8.1368086       1.644854 NA
#> resample_144  0.05 0.9999831 1.0000000     ns     -8.0848079       1.644854 NA
#> resample_145  0.05 0.9999831 1.0000000     ns     -8.0958262       1.644854 NA
#> resample_146  0.05 0.9999832 1.0000000     ns     -8.1001894       1.644854 NA
#> resample_147  0.05 0.9999836 1.0000000     ns     -8.1728657       1.644854 NA
#> resample_148  0.05 0.9999834 1.0000000     ns     -8.1417958       1.644854 NA
#> resample_149  0.05 0.9999831 1.0000000     ns     -8.0964257       1.644854 NA
#> resample_150  0.05 0.9999831 1.0000000     ns     -8.0886733       1.644854 NA
#> resample_151  0.05 0.9999829 1.0000000     ns     -8.0573797       1.644854 NA
#> resample_152  0.05 0.9999827 1.0000000     ns     -8.0297190       1.644854 NA
#> resample_153  0.05 0.9999824 1.0000000     ns     -7.9898096       1.644854 NA
#> resample_154  0.05 0.9999824 1.0000000     ns     -7.9859556       1.644854 NA
#> resample_155  0.05 0.9999823 1.0000000     ns     -7.9738089       1.644854 NA
#> resample_156  0.05 0.9999823 1.0000000     ns     -7.9657107       1.644854 NA
#> resample_157  0.05 0.9999819 1.0000000     ns     -7.9061372       1.644854 NA
#> resample_158  0.05 0.9999815 1.0000000     ns     -7.8529784       1.644854 NA
#> resample_159  0.05 0.9999814 1.0000000     ns     -7.8338235       1.644854 NA
#> resample_160  0.05 0.9999811 1.0000000     ns     -7.7983677       1.644854 NA
#> resample_161  0.05 0.9999808 1.0000000     ns     -7.7510574       1.644854 NA
#> resample_162  0.05 0.9999804 1.0000000     ns     -7.7070002       1.644854 NA
#> resample_163  0.05 0.9999801 1.0000000     ns     -7.6565703       1.644854 NA
#> resample_164  0.05 0.9999797 1.0000000     ns     -7.6097553       1.644854 NA
#> resample_165  0.05 0.9999794 1.0000000     ns     -7.5715629       1.644854 NA
#> resample_166  0.05 0.9999793 1.0000000     ns     -7.5572966       1.644854 NA
#> resample_167  0.05 0.9999790 1.0000000     ns     -7.5171108       1.644854 NA
#> resample_168  0.05 0.9999787 1.0000000     ns     -7.4786221       1.644854 NA
#> resample_169  0.05 0.9999786 1.0000000     ns     -7.4710825       1.644854 NA
#> resample_170  0.05 0.9999782 1.0000000     ns     -7.4182155       1.644854 NA
#> resample_171  0.05 0.9999777 1.0000000     ns     -7.3580214       1.644854 NA
#> resample_172  0.05 0.9999774 1.0000000     ns     -7.3308900       1.644854 NA
#> resample_173  0.05 0.9999770 1.0000000     ns     -7.2830657       1.644854 NA
#> resample_174  0.05 0.9999767 1.0000000     ns     -7.2440626       1.644854 NA
#> resample_175  0.05 0.9999763 1.0000000     ns     -7.2026683       1.644854 NA
#> resample_176  0.05 0.9999759 1.0000000     ns     -7.1590737       1.644854 NA
#> resample_177  0.05 0.9999755 1.0000000     ns     -7.1213584       1.644854 NA
#> resample_178  0.05 0.9999748 1.0000000     ns     -7.0435234       1.644854 NA
#> resample_179  0.05 0.9999743 1.0000000     ns     -6.9872665       1.644854 NA
#> resample_180  0.05 0.9999728 1.0000000     ns     -6.8471977       1.644854 NA
#> resample_181  0.05 0.9999721 1.0000000     ns     -6.7802393       1.644854 NA
#> resample_182  0.05 0.9999717 1.0000000     ns     -6.7377833       1.644854 NA
#> resample_183  0.05 0.9999710 1.0000000     ns     -6.6761075       1.644854 NA
#> resample_184  0.05 0.9999703 1.0000000     ns     -6.6158383       1.644854 NA
#> resample_185  0.05 0.9999675 1.0000000     ns     -6.3731797       1.644854 NA
#> resample_186  0.05 0.9999659 1.0000000     ns     -6.2523587       1.644854 NA
#> resample_187  0.05 0.9999654 1.0000000     ns     -6.2085808       1.644854 NA
#> resample_188  0.05 0.9999650 1.0000000     ns     -6.1833462       1.644854 NA
#> resample_189  0.05 0.9999642 1.0000000     ns     -6.1185571       1.644854 NA
#> resample_190  0.05 0.9999637 1.0000000     ns     -6.0826999       1.644854 NA
#> resample_191  0.05 0.9999630 1.0000000     ns     -6.0373649       1.644854 NA
#> resample_192  0.05 0.9999625 1.0000000     ns     -5.9967689       1.644854 NA
#> resample_193  0.05 0.9999534 1.0000000     ns     -5.4295390       1.644854 NA
#> resample_194  0.05 0.9999499 0.9999999     ns     -5.2393488       1.644854 NA
#> resample_195  0.05 0.9999455 0.9999997     ns     -5.0212721       1.644854 NA
#> resample_196  0.05 0.9999401 0.9999991     ns     -4.7707037       1.644854 NA
#> resample_197  0.05 0.9999296 0.9999931     ns     -4.3479697       1.644854 NA
#> resample_198  0.05 0.9999273 0.9999900     ns     -4.2638316       1.644854 NA
#> resample_199  0.05 0.9999010 0.9997212     ns     -3.4514508       1.644854 NA
#> resample_200  0.05 0.9998529 0.9921169     ns     -2.4142846       1.644854 NA
#>              CI_lower CI_upper    N           AVG          MED       MIN
#> resample_1         NA       NA 5000 -1.205872e-03 -0.001743506 -5.339639
#> resample_2         NA       NA 5000 -1.187793e-03 -0.001740522 -5.287884
#> resample_3         NA       NA 5000 -1.063386e-03 -0.001739585 -4.702157
#> resample_4         NA       NA 5000 -1.037249e-03 -0.001737876 -4.608625
#> resample_5         NA       NA 5000 -1.026229e-03 -0.001735504 -4.589538
#> resample_6         NA       NA 5000 -1.009276e-03 -0.001730487 -4.541207
#> resample_7         NA       NA 5000 -9.922198e-04 -0.001728383 -4.491477
#> resample_8         NA       NA 5000 -9.712276e-04 -0.001725680 -4.425886
#> resample_9         NA       NA 5000 -9.627666e-04 -0.001724486 -4.418752
#> resample_10        NA       NA 5000 -9.413915e-04 -0.001718344 -4.345208
#> resample_11        NA       NA 5000 -9.304777e-04 -0.001716193 -4.328513
#> resample_12        NA       NA 5000 -9.199843e-04 -0.001715437 -4.309326
#> resample_13        NA       NA 5000 -9.069660e-04 -0.001712890 -4.283795
#> resample_14        NA       NA 5000 -8.982599e-04 -0.001712518 -4.282030
#> resample_15        NA       NA 5000 -8.907250e-04 -0.001710912 -4.280437
#> resample_16        NA       NA 5000 -8.816850e-04 -0.001704195 -4.273743
#> resample_17        NA       NA 5000 -8.629269e-04 -0.001703191 -4.220920
#> resample_18        NA       NA 5000 -8.542176e-04 -0.001701940 -4.210686
#> resample_19        NA       NA 5000 -8.471323e-04 -0.001700468 -4.210024
#> resample_20        NA       NA 5000 -8.333293e-04 -0.001698916 -4.176450
#> resample_21        NA       NA 5000 -8.247428e-04 -0.001698274 -4.167878
#> resample_22        NA       NA 5000 -8.144613e-04 -0.001697654 -4.150007
#> resample_23        NA       NA 5000 -8.063466e-04 -0.001694535 -4.147907
#> resample_24        NA       NA 5000 -7.943123e-04 -0.001693288 -4.124231
#> resample_25        NA       NA 5000 -7.837763e-04 -0.001693122 -4.108253
#> resample_26        NA       NA 5000 -7.721291e-04 -0.001688223 -4.083190
#> resample_27        NA       NA 5000 -7.643056e-04 -0.001688034 -4.079785
#> resample_28        NA       NA 5000 -7.567237e-04 -0.001687487 -4.077277
#> resample_29        NA       NA 5000 -7.491102e-04 -0.001685377 -4.073478
#> resample_30        NA       NA 5000 -7.409426e-04 -0.001684423 -4.068973
#> resample_31        NA       NA 5000 -7.333125e-04 -0.001680015 -4.066269
#> resample_32        NA       NA 5000 -7.255608e-04 -0.001676304 -4.062316
#> resample_33        NA       NA 5000 -7.171450e-04 -0.001674515 -4.056479
#> resample_34        NA       NA 5000 -7.063857e-04 -0.001673473 -4.039766
#> resample_35        NA       NA 5000 -6.980131e-04 -0.001670031 -4.037651
#> resample_36        NA       NA 5000 -6.906337e-04 -0.001668594 -4.035776
#> resample_37        NA       NA 5000 -6.822440e-04 -0.001663825 -4.032000
#> resample_38        NA       NA 5000 -6.743398e-04 -0.001660392 -4.028427
#> resample_39        NA       NA 5000 -6.663693e-04 -0.001650246 -4.023787
#> resample_40        NA       NA 5000 -6.552093e-04 -0.001648426 -4.004583
#> resample_41        NA       NA 5000 -6.478683e-04 -0.001644579 -4.001214
#> resample_42        NA       NA 5000 -6.398952e-04 -0.001642286 -3.997698
#> resample_43        NA       NA 5000 -6.309221e-04 -0.001640811 -3.990283
#> resample_44        NA       NA 5000 -6.223083e-04 -0.001635816 -3.989083
#> resample_45        NA       NA 5000 -6.109675e-04 -0.001634608 -3.968141
#> resample_46        NA       NA 5000 -6.004822e-04 -0.001634038 -3.955319
#> resample_47        NA       NA 5000 -5.923178e-04 -0.001630752 -3.955113
#> resample_48        NA       NA 5000 -5.818929e-04 -0.001627185 -3.939586
#> resample_49        NA       NA 5000 -5.737097e-04 -0.001625811 -3.933637
#> resample_50        NA       NA 5000 -5.654923e-04 -0.001623479 -3.929024
#> resample_51        NA       NA 5000 -5.582764e-04 -0.001622063 -3.927118
#> resample_52        NA       NA 5000 -5.512165e-04 -0.001620875 -3.927117
#> resample_53        NA       NA 5000 -5.421367e-04 -0.001620427 -3.920254
#> resample_54        NA       NA 5000 -5.345757e-04 -0.001617866 -3.920009
#> resample_55        NA       NA 5000 -5.267425e-04 -0.001617381 -3.917952
#> resample_56        NA       NA 5000 -5.188605e-04 -0.001610533 -3.916937
#> resample_57        NA       NA 5000 -5.079181e-04 -0.001608846 -3.899840
#> resample_58        NA       NA 5000 -4.972451e-04 -0.001607131 -3.881852
#> resample_59        NA       NA 5000 -4.894138e-04 -0.001605020 -3.874809
#> resample_60        NA       NA 5000 -4.813613e-04 -0.001601585 -3.871397
#> resample_61        NA       NA 5000 -4.733975e-04 -0.001599551 -3.865413
#> resample_62        NA       NA 5000 -4.652166e-04 -0.001596200 -3.865099
#> resample_63        NA       NA 5000 -4.574845e-04 -0.001591517 -3.862207
#> resample_64        NA       NA 5000 -4.496466e-04 -0.001590370 -3.857484
#> resample_65        NA       NA 5000 -4.401492e-04 -0.001585100 -3.854791
#> resample_66        NA       NA 5000 -4.318272e-04 -0.001580646 -3.852338
#> resample_67        NA       NA 5000 -4.238186e-04 -0.001578021 -3.849123
#> resample_68        NA       NA 5000 -4.161195e-04 -0.001574843 -3.847579
#> resample_69        NA       NA 5000 -4.079401e-04 -0.001574484 -3.845277
#> resample_70        NA       NA 5000 -4.006767e-04 -0.001572993 -3.844185
#> resample_71        NA       NA 5000 -3.919035e-04 -0.001567502 -3.841991
#> resample_72        NA       NA 5000 -3.833688e-04 -0.001564550 -3.835498
#> resample_73        NA       NA 5000 -3.759191e-04 -0.001560487 -3.835338
#> resample_74        NA       NA 5000 -3.666883e-04 -0.001557283 -3.832666
#> resample_75        NA       NA 5000 -3.594123e-04 -0.001557118 -3.831857
#> resample_76        NA       NA 5000 -3.519357e-04 -0.001554189 -3.830687
#> resample_77        NA       NA 5000 -3.445599e-04 -0.001553668 -3.829363
#> resample_78        NA       NA 5000 -3.369225e-04 -0.001552850 -3.829153
#> resample_79        NA       NA 5000 -3.273414e-04 -0.001552086 -3.815205
#> resample_80        NA       NA 5000 -3.201268e-04 -0.001550587 -3.815023
#> resample_81        NA       NA 5000 -3.117997e-04 -0.001548485 -3.810992
#> resample_82        NA       NA 5000 -3.042427e-04 -0.001544094 -3.809532
#> resample_83        NA       NA 5000 -2.973632e-04 -0.001543072 -3.807995
#> resample_84        NA       NA 5000 -2.898073e-04 -0.001541031 -3.807751
#> resample_85        NA       NA 5000 -2.821636e-04 -0.001535346 -3.807136
#> resample_86        NA       NA 5000 -2.747204e-04 -0.001533630 -3.806807
#> resample_87        NA       NA 5000 -2.666454e-04 -0.001532205 -3.802386
#> resample_88        NA       NA 5000 -2.592413e-04 -0.001531020 -3.802348
#> resample_89        NA       NA 5000 -2.520585e-04 -0.001530132 -3.801098
#> resample_90        NA       NA 5000 -2.425888e-04 -0.001529275 -3.799154
#> resample_91        NA       NA 5000 -2.345104e-04 -0.001528476 -3.797727
#> resample_92        NA       NA 5000 -2.262155e-04 -0.001525439 -3.796379
#> resample_93        NA       NA 5000 -2.187871e-04 -0.001522916 -3.794229
#> resample_94        NA       NA 5000 -2.085684e-04 -0.001520232 -3.779512
#> resample_95        NA       NA 5000 -1.997300e-04 -0.001514947 -3.772302
#> resample_96        NA       NA 5000 -1.896541e-04 -0.001512817 -3.764507
#> resample_97        NA       NA 5000 -1.815001e-04 -0.001508281 -3.763452
#> resample_98        NA       NA 5000 -1.724876e-04 -0.001501603 -3.761226
#> resample_99        NA       NA 5000 -1.655592e-04 -0.001499325 -3.761198
#> resample_100       NA       NA 5000 -1.580916e-04 -0.001495488 -3.760663
#> resample_101       NA       NA 5000 -1.499939e-04 -0.001492029 -3.757176
#> resample_102       NA       NA 5000 -1.416769e-04 -0.001489979 -3.750920
#> resample_103       NA       NA 5000 -1.331819e-04 -0.001487774 -3.743326
#> resample_104       NA       NA 5000 -1.247951e-04 -0.001485401 -3.735419
#> resample_105       NA       NA 5000 -1.153504e-04 -0.001482331 -3.727975
#> resample_106       NA       NA 5000 -1.083209e-04 -0.001480667 -3.727017
#> resample_107       NA       NA 5000 -9.977331e-05 -0.001477653 -3.725370
#> resample_108       NA       NA 5000 -9.274265e-05 -0.001476636 -3.725050
#> resample_109       NA       NA 5000 -8.503779e-05 -0.001473341 -3.724896
#> resample_110       NA       NA 5000 -7.773570e-05 -0.001472063 -3.723011
#> resample_111       NA       NA 5000 -6.954828e-05 -0.001468972 -3.721701
#> resample_112       NA       NA 5000 -6.192408e-05 -0.001465304 -3.720658
#> resample_113       NA       NA 5000 -5.496089e-05 -0.001464507 -3.718167
#> resample_114       NA       NA 5000 -4.550415e-05 -0.001463374 -3.707864
#> resample_115       NA       NA 5000 -3.727543e-05 -0.001460367 -3.707690
#> resample_116       NA       NA 5000 -2.876078e-05 -0.001458646 -3.707623
#> resample_117       NA       NA 5000 -1.911859e-05 -0.001456935 -3.701323
#> resample_118       NA       NA 5000 -1.110999e-05 -0.001453581 -3.700625
#> resample_119       NA       NA 5000 -3.778834e-06 -0.001452069 -3.698825
#> resample_120       NA       NA 5000  5.189091e-06 -0.001450732 -3.694647
#> resample_121       NA       NA 5000  1.279639e-05 -0.001446487 -3.693574
#> resample_122       NA       NA 5000  2.005132e-05 -0.001445870 -3.692502
#> resample_123       NA       NA 5000  2.871716e-05 -0.001442429 -3.689411
#> resample_124       NA       NA 5000  3.625029e-05 -0.001442095 -3.688343
#> resample_125       NA       NA 5000  4.310733e-05 -0.001437804 -3.687853
#> resample_126       NA       NA 5000  5.229409e-05 -0.001435442 -3.687809
#> resample_127       NA       NA 5000  5.996368e-05 -0.001433941 -3.687274
#> resample_128       NA       NA 5000  6.868755e-05 -0.001429819 -3.684233
#> resample_129       NA       NA 5000  7.629953e-05 -0.001428407 -3.680632
#> resample_130       NA       NA 5000  8.484073e-05 -0.001426085 -3.675784
#> resample_131       NA       NA 5000  9.372990e-05 -0.001425687 -3.674472
#> resample_132       NA       NA 5000  1.006387e-04 -0.001424151 -3.674144
#> resample_133       NA       NA 5000  1.085544e-04 -0.001421322 -3.671407
#> resample_134       NA       NA 5000  1.165249e-04 -0.001419725 -3.669956
#> resample_135       NA       NA 5000  1.249613e-04 -0.001418910 -3.666922
#> resample_136       NA       NA 5000  1.321746e-04 -0.001414369 -3.666314
#> resample_137       NA       NA 5000  1.397946e-04 -0.001409314 -3.665150
#> resample_138       NA       NA 5000  1.476034e-04 -0.001407840 -3.663860
#> resample_139       NA       NA 5000  1.553742e-04 -0.001405144 -3.662064
#> resample_140       NA       NA 5000  1.646553e-04 -0.001403780 -3.657497
#> resample_141       NA       NA 5000  1.738675e-04 -0.001401650 -3.654982
#> resample_142       NA       NA 5000  1.819775e-04 -0.001398937 -3.652887
#> resample_143       NA       NA 5000  1.905031e-04 -0.001397459 -3.647896
#> resample_144       NA       NA 5000  1.991875e-04 -0.001396906 -3.643731
#> resample_145       NA       NA 5000  2.069455e-04 -0.001395503 -3.643651
#> resample_146       NA       NA 5000  2.156617e-04 -0.001391910 -3.642542
#> resample_147       NA       NA 5000  2.271806e-04 -0.001390404 -3.642248
#> resample_148       NA       NA 5000  2.355551e-04 -0.001389128 -3.640131
#> resample_149       NA       NA 5000  2.433318e-04 -0.001387000 -3.637483
#> resample_150       NA       NA 5000  2.509436e-04 -0.001386395 -3.637109
#> resample_151       NA       NA 5000  2.607065e-04 -0.001384911 -3.633214
#> resample_152       NA       NA 5000  2.681589e-04 -0.001381920 -3.631911
#> resample_153       NA       NA 5000  2.789926e-04 -0.001381289 -3.628634
#> resample_154       NA       NA 5000  2.874335e-04 -0.001377866 -3.628351
#> resample_155       NA       NA 5000  2.971537e-04 -0.001376715 -3.627833
#> resample_156       NA       NA 5000  3.055047e-04 -0.001375034 -3.627789
#> resample_157       NA       NA 5000  3.146325e-04 -0.001374564 -3.624512
#> resample_158       NA       NA 5000  3.224136e-04 -0.001372543 -3.621615
#> resample_159       NA       NA 5000  3.304979e-04 -0.001370680 -3.621226
#> resample_160       NA       NA 5000  3.393662e-04 -0.001367288 -3.620233
#> resample_161       NA       NA 5000  3.479925e-04 -0.001365743 -3.618350
#> resample_162       NA       NA 5000  3.561119e-04 -0.001363444 -3.616621
#> resample_163       NA       NA 5000  3.653156e-04 -0.001362706 -3.614819
#> resample_164       NA       NA 5000  3.733138e-04 -0.001360507 -3.612490
#> resample_165       NA       NA 5000  3.828222e-04 -0.001355583 -3.612296
#> resample_166       NA       NA 5000  3.898202e-04 -0.001351761 -3.612086
#> resample_167       NA       NA 5000  3.969404e-04 -0.001350262 -3.609975
#> resample_168       NA       NA 5000  4.050640e-04 -0.001346873 -3.608712
#> resample_169       NA       NA 5000  4.116489e-04 -0.001345719 -3.608660
#> resample_170       NA       NA 5000  4.217797e-04 -0.001344683 -3.608249
#> resample_171       NA       NA 5000  4.307005e-04 -0.001344246 -3.605063
#> resample_172       NA       NA 5000  4.380499e-04 -0.001339979 -3.603769
#> resample_173       NA       NA 5000  4.479512e-04 -0.001335687 -3.603745
#> resample_174       NA       NA 5000  4.563600e-04 -0.001331332 -3.603540
#> resample_175       NA       NA 5000  4.643791e-04 -0.001327076 -3.602696
#> resample_176       NA       NA 5000  4.728976e-04 -0.001324926 -3.600904
#> resample_177       NA       NA 5000  4.806396e-04 -0.001324120 -3.598943
#> resample_178       NA       NA 5000  4.899969e-04 -0.001323569 -3.596994
#> resample_179       NA       NA 5000  4.991258e-04 -0.001318850 -3.596608
#> resample_180       NA       NA 5000  5.106086e-04 -0.001317074 -3.592841
#> resample_181       NA       NA 5000  5.189131e-04 -0.001313950 -3.590578
#> resample_182       NA       NA 5000  5.270497e-04 -0.001309357 -3.589893
#> resample_183       NA       NA 5000  5.358593e-04 -0.001306051 -3.588360
#> resample_184       NA       NA 5000  5.447057e-04 -0.001305495 -3.586814
#> resample_185       NA       NA 5000  5.594344e-04 -0.001300308 -3.586183
#> resample_186       NA       NA 5000  5.697645e-04 -0.001298727 -3.584759
#> resample_187       NA       NA 5000  5.775272e-04 -0.001298392 -3.581442
#> resample_188       NA       NA 5000  5.849053e-04 -0.001296476 -3.581041
#> resample_189       NA       NA 5000  5.938718e-04 -0.001292676 -3.580128
#> resample_190       NA       NA 5000  6.017235e-04 -0.001291121 -3.579091
#> resample_191       NA       NA 5000  6.093491e-04 -0.001289051 -3.575704
#> resample_192       NA       NA 5000  6.174868e-04 -0.001282788 -3.573042
#> resample_193       NA       NA 5000  6.408088e-04 -0.001280618 -3.572062
#> resample_194       NA       NA 5000  6.532673e-04 -0.001276494 -3.570103
#> resample_195       NA       NA 5000  6.664166e-04 -0.001275460 -3.567820
#> resample_196       NA       NA 5000  6.801835e-04 -0.001273583 -3.566900
#> resample_197       NA       NA 5000  6.999353e-04 -0.001272958 -3.565921
#> resample_198       NA       NA 5000  7.101414e-04 -0.001271878 -3.564643
#> resample_199       NA       NA 5000  7.416341e-04 -0.001268941 -3.563582
#> resample_200       NA       NA 5000  7.853101e-04 -0.001265717 -3.562775
#>                   MAX        SD
#> resample_1   3.532265 1.0012018
#> resample_2   3.533722 1.0011463
#> resample_3   3.534160 1.0005624
#> resample_4   3.534343 1.0004770
#> resample_5   3.534606 1.0004568
#> resample_6   3.536013 1.0004124
#> resample_7   3.536026 1.0003682
#> resample_8   3.536708 1.0003094
#> resample_9   3.537325 1.0003020
#> resample_10  3.537701 1.0002367
#> resample_11  3.542142 1.0002254
#> resample_12  3.542353 1.0002074
#> resample_13  3.542605 1.0001808
#> resample_14  3.547840 1.0001814
#> resample_15  3.548945 1.0001800
#> resample_16  3.549228 1.0001740
#> resample_17  3.550101 1.0001271
#> resample_18  3.550267 1.0001182
#> resample_19  3.550916 1.0001178
#> resample_20  3.551506 1.0000896
#> resample_21  3.551640 1.0000813
#> resample_22  3.551713 1.0000660
#> resample_23  3.553910 1.0000633
#> resample_24  3.556550 1.0000461
#> resample_25  3.556967 1.0000314
#> resample_26  3.557575 1.0000108
#> resample_27  3.557753 1.0000076
#> resample_28  3.559957 1.0000070
#> resample_29  3.561713 1.0000042
#> resample_30  3.561932 1.0000001
#> resample_31  3.562437 0.9999962
#> resample_32  3.562725 0.9999927
#> resample_33  3.564624 0.9999895
#> resample_34  3.566401 0.9999770
#> resample_35  3.569961 0.9999765
#> resample_36  3.571166 0.9999743
#> resample_37  3.571685 0.9999701
#> resample_38  3.572697 0.9999670
#> resample_39  3.575345 0.9999670
#> resample_40  3.577162 0.9999517
#> resample_41  3.577903 0.9999490
#> resample_42  3.579043 0.9999466
#> resample_43  3.579578 0.9999402
#> resample_44  3.581528 0.9999351
#> resample_45  3.583458 0.9999205
#> resample_46  3.586945 0.9999122
#> resample_47  3.587866 0.9999162
#> resample_48  3.588949 0.9999058
#> resample_49  3.589337 0.9999034
#> resample_50  3.589918 0.9999004
#> resample_51  3.590616 0.9998986
#> resample_52  3.591495 0.9998993
#> resample_53  3.595614 0.9998967
#> resample_54  3.596912 0.9998962
#> resample_55  3.598600 0.9998952
#> resample_56  3.599231 0.9998949
#> resample_57  3.603651 0.9998849
#> resample_58  3.604178 0.9998708
#> resample_59  3.604209 0.9998651
#> resample_60  3.604494 0.9998590
#> resample_61  3.605579 0.9998549
#> resample_62  3.608577 0.9998551
#> resample_63  3.610418 0.9998538
#> resample_64  3.610898 0.9998495
#> resample_65  3.613702 0.9998428
#> resample_66  3.615369 0.9998398
#> resample_67  3.616844 0.9998369
#> resample_68  3.619244 0.9998368
#> resample_69  3.620297 0.9998373
#> resample_70  3.620590 0.9998360
#> resample_71  3.626649 0.9998391
#> resample_72  3.628571 0.9998357
#> resample_73  3.632271 0.9998389
#> resample_74  3.636804 0.9998391
#> resample_75  3.637524 0.9998380
#> resample_76  3.638179 0.9998353
#> resample_77  3.639766 0.9998356
#> resample_78  3.641782 0.9998369
#> resample_79  3.642386 0.9998252
#> resample_80  3.643045 0.9998266
#> resample_81  3.645305 0.9998266
#> resample_82  3.647813 0.9998299
#> resample_83  3.648848 0.9998293
#> resample_84  3.649571 0.9998294
#> resample_85  3.651394 0.9998300
#> resample_86  3.652160 0.9998300
#> resample_87  3.656134 0.9998292
#> resample_88  3.657682 0.9998317
#> resample_89  3.658426 0.9998312
#> resample_90  3.664307 0.9998347
#> resample_91  3.665790 0.9998333
#> resample_92  3.669785 0.9998351
#> resample_93  3.670907 0.9998332
#> resample_94  3.671358 0.9998232
#> resample_95  3.674161 0.9998195
#> resample_96  3.677748 0.9998148
#> resample_97  3.681549 0.9998177
#> resample_98  3.687539 0.9998178
#> resample_99  3.690134 0.9998188
#> resample_100 3.690551 0.9998199
#> resample_101 3.692214 0.9998197
#> resample_102 3.693299 0.9998154
#> resample_103 3.695537 0.9998109
#> resample_104 3.695600 0.9998045
#> resample_105 3.695997 0.9997984
#> resample_106 3.696300 0.9997979
#> resample_107 3.699728 0.9997999
#> resample_108 3.700158 0.9998005
#> resample_109 3.702752 0.9998038
#> resample_110 3.703228 0.9998026
#> resample_111 3.706334 0.9998049
#> resample_112 3.706727 0.9998045
#> resample_113 3.706761 0.9998022
#> resample_114 3.707036 0.9997961
#> resample_115 3.709133 0.9998005
#> resample_116 3.715700 0.9998067
#> resample_117 3.724249 0.9998094
#> resample_118 3.729282 0.9998133
#> resample_119 3.729920 0.9998131
#> resample_120 3.731063 0.9998077
#> resample_121 3.734342 0.9998094
#> resample_122 3.737682 0.9998111
#> resample_123 3.740787 0.9998103
#> resample_124 3.742260 0.9998107
#> resample_125 3.743013 0.9998111
#> resample_126 3.746575 0.9998120
#> resample_127 3.750588 0.9998151
#> resample_128 3.756782 0.9998179
#> resample_129 3.756960 0.9998162
#> resample_130 3.758178 0.9998120
#> resample_131 3.763408 0.9998139
#> resample_132 3.763837 0.9998140
#> resample_133 3.766847 0.9998144
#> resample_134 3.768870 0.9998141
#> resample_135 3.772150 0.9998147
#> resample_136 3.774238 0.9998152
#> resample_137 3.778209 0.9998169
#> resample_138 3.781716 0.9998190
#> resample_139 3.781742 0.9998198
#> resample_140 3.791369 0.9998241
#> resample_141 3.794788 0.9998231
#> resample_142 3.797615 0.9998241
#> resample_143 3.798256 0.9998224
#> resample_144 3.801254 0.9998223
#> resample_145 3.804643 0.9998248
#> resample_146 3.811467 0.9998298
#> resample_147 3.832832 0.9998451
#> resample_148 3.835902 0.9998465
#> resample_149 3.836128 0.9998449
#> resample_150 3.837368 0.9998458
#> resample_151 3.850100 0.9998526
#> resample_152 3.850339 0.9998530
#> resample_153 3.865182 0.9998599
#> resample_154 3.874152 0.9998679
#> resample_155 3.884615 0.9998743
#> resample_156 3.891571 0.9998794
#> resample_157 3.898250 0.9998824
#> resample_158 3.898648 0.9998826
#> resample_159 3.904620 0.9998863
#> resample_160 3.908549 0.9998919
#> resample_161 3.916502 0.9998952
#> resample_162 3.917779 0.9998983
#> resample_163 3.926112 0.9999035
#> resample_164 3.926603 0.9999060
#> resample_165 3.933528 0.9999067
#> resample_166 3.934811 0.9999097
#> resample_167 3.936950 0.9999106
#> resample_168 3.941538 0.9999120
#> resample_169 3.941853 0.9999110
#> resample_170 3.952013 0.9999194
#> resample_171 3.954234 0.9999174
#> resample_172 3.954644 0.9999167
#> resample_173 3.965079 0.9999274
#> resample_174 3.972897 0.9999340
#> resample_175 3.978307 0.9999376
#> resample_176 3.981288 0.9999382
#> resample_177 3.982624 0.9999379
#> resample_178 3.993590 0.9999449
#> resample_179 4.002394 0.9999504
#> resample_180 4.019563 0.9999630
#> resample_181 4.026229 0.9999666
#> resample_182 4.031483 0.9999707
#> resample_183 4.036868 0.9999766
#> resample_184 4.042782 0.9999821
#> resample_185 4.080379 1.0000126
#> resample_186 4.095388 1.0000273
#> resample_187 4.096718 1.0000263
#> resample_188 4.099363 1.0000289
#> resample_189 4.106527 1.0000347
#> resample_190 4.109845 1.0000371
#> resample_191 4.110795 1.0000377
#> resample_192 4.112246 1.0000385
#> resample_193 4.194489 1.0001052
#> resample_194 4.218670 1.0001246
#> resample_195 4.247214 1.0001479
#> resample_196 4.281704 1.0001782
#> resample_197 4.341693 1.0002293
#> resample_198 4.352137 1.0002408
#> resample_199 4.475621 1.0003506
#> resample_200 4.651232 1.0005141